requirements.txt
PyQt6==6.6.1
textblob==0.18.0
nltk==3.8.1
pycryptodome==3.20.0